Create thumb_twiddle.sh
The embedded SOM that I'm working on doesn't have a very good /proc/cpuinfo, so here's a quick script to give me a very rough estimate of their CPU power. My initial use case was to compare an old SOM to a newer one to see if they were at different clock speeds.